Move more classes into their own files
